refactor(frontend): extract exposed-ports editor modal from main.ts (#312)
Continues the conservative split: pull the exposed-ports editor (#190 —
live-edit a session's declared config.ports[]) into ports.ts (~252 lines).

Same TDZ-safe pattern: DOM re-queried locally; main.ts imports (showToast,
activeSessionId, sessions) used only inside functions; the shared portsModal
const stays in main.ts for the Escape guard; main.ts imports closePortsModal
for its modal-registry callback.

main.ts: 4070 → 3831. tsc + Biome clean, vite build green, frontend tests
pass (66).
